+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#741** : Added a count unique dashboard table expression function.
+
 ## [v5.2.4] - 2018-04-18
 
 * Issue **#735** : Change the format-date xslt function to parse dates in a case insensitive way.